Move CLI files into cli sub-module
