+++ tests/cli/cli_gen_output.py
