+++ tests/cli/test_cli_gen.py
@@ -9,29 +9,29 @@ from unittest import mock
 import pytest
 import typer
 
-from openapi_spec_tools.cli_gen.cli import TreeDisplay
-from openapi_spec_tools.cli_gen.cli import TreeFormat
-from openapi_spec_tools.cli_gen.cli import generate_check_missing
-from openapi_spec_tools.cli_gen.cli import generate_cli
-from openapi_spec_tools.cli_gen.cli import generate_unreferenced
-from openapi_spec_tools.cli_gen.cli import layout_check_format
-from openapi_spec_tools.cli_gen.cli import layout_operations
-from openapi_spec_tools.cli_gen.cli import layout_suggest
-from openapi_spec_tools.cli_gen.cli import layout_tree
-from openapi_spec_tools.cli_gen.cli import layout_tree_with_error_handling
-from openapi_spec_tools.cli_gen.cli import open_layout_with_error_handling
-from openapi_spec_tools.cli_gen.cli import open_oas_with_error_handling
-from openapi_spec_tools.cli_gen.cli import show_cli_tree
-from openapi_spec_tools.cli_gen.cli import trim_oas
-from tests.cli_gen.cli_output import P_V_ALL
-from tests.cli_gen.cli_output import P_V_MID
-from tests.cli_gen.cli_output import P_V_PETS
-from tests.cli_gen.cli_output import P_V_SHALLOW
-from tests.cli_gen.cli_output import PET_ALL
-from tests.cli_gen.cli_output import PET_FUNC
-from tests.cli_gen.cli_output import PET_HELP
-from tests.cli_gen.cli_output import PET_OP
-from tests.cli_gen.cli_output import PET_PATH
+from openapi_spec_tools.cli.cli_gen import TreeDisplay
+from openapi_spec_tools.cli.cli_gen import TreeFormat
+from openapi_spec_tools.cli.cli_gen import generate_check_missing
+from openapi_spec_tools.cli.cli_gen import generate_cli
+from openapi_spec_tools.cli.cli_gen import generate_unreferenced
+from openapi_spec_tools.cli.cli_gen import layout_check_format
+from openapi_spec_tools.cli.cli_gen import layout_operations
+from openapi_spec_tools.cli.cli_gen import layout_suggest
+from openapi_spec_tools.cli.cli_gen import layout_tree
+from openapi_spec_tools.cli.cli_gen import layout_tree_with_error_handling
+from openapi_spec_tools.cli.cli_gen import open_layout_with_error_handling
+from openapi_spec_tools.cli.cli_gen import open_oas_with_error_handling
+from openapi_spec_tools.cli.cli_gen import show_cli_tree
+from openapi_spec_tools.cli.cli_gen import trim_oas
+from tests.cli.cli_gen_output import P_V_ALL
+from tests.cli.cli_gen_output import P_V_MID
+from tests.cli.cli_gen_output import P_V_PETS
+from tests.cli.cli_gen_output import P_V_SHALLOW
+from tests.cli.cli_gen_output import PET_ALL
+from tests.cli.cli_gen_output import PET_FUNC
+from tests.cli.cli_gen_output import PET_HELP
+from tests.cli.cli_gen_output import PET_OP
+from tests.cli.cli_gen_output import PET_PATH
 from tests.cli_gen.helpers import to_ascii
 from tests.helpers import StringIo
 from tests.helpers import asset_filename
